+++ ExpandWorldPrefabs/Commands.cs
@@ -13,7 +13,7 @@ public class Commands
   public static void Run(Info info, Parameters pars)
   {
     if (info.Commands.Length == 0) return;
-    var commands = info.Commands.Select(pars.Replace).ToArray();
+    var commands = info.Commands.Select(c => pars.Replace(c, true)).ToArray();
     Run(commands);
   }